+++ CHANGELOG.md
@@ -28,6 +28,8 @@ All notable changes to this project will be documented in this file, in reverse 
 - [#48](https://github.com/webimpress/coding-standard/pull/48) adds configuration option `nullPosition` to `Commenting\TagWithType` sniff. Default value is `first` to keep backward compatibility.
   The other allowed value is `last` so then `null` values in type list is at the last position.
 
+- [#49](https://github.com/webimpress/coding-standard/pull/49) adds `Commenting\DisallowEmptyComment` sniff to detect empty comments and multiple empty lines in comments
+
 ### Changed
 
 - [#42](https://github.com/webimpress/coding-standard/pull/42) changes `NamingConventions\ValidVariableName` to require variable names be in strict camelCase. It means two capital letters next to each other are not allowed.